refactor(gltf_auto_export): continued restructure & cleanup  * removed redundant preferences (once in operator once seperatly)  * simplified logic of operator to account for new changes  * modified tracker to work with changed operator  * added special flag to operator to be able to call it directly & still have the correct settings  * updates to ui & helpers  * renamed stuff for clarity  * various related code migrations
