+++ tools/gltf_auto_export/__init__.py
@@ -15,7 +15,8 @@ import os
 
 from bpy.props import (IntProperty)
 
-from .tracker.tracker import AutoExportTracker
+from .auto_export.operators import AutoExportGLTF
+from .auto_export.tracker import AutoExportTracker
 from . import helpers
 from .internals import (SceneLink,
                         SceneLinks,
@@ -30,10 +31,8 @@ from .ui.main import (GLTF_PT_auto_export_main,
                       GLTF_PT_auto_export_collections_list,
                       GLTF_PT_auto_export_gltf,
                       SCENE_UL_GLTF_auto_export,
-                      AutoExportGLTF
                       )
-from .ui.various import (SCENES_LIST_OT_actions)
-from .helpers_scenes import (is_scene_ok)
+from .ui.operators import (SCENES_LIST_OT_actions)
 
 
 ######################################################
@@ -74,31 +73,13 @@ def register():
     for cls in classes:
         bpy.utils.register_class(cls)
 
-    bpy.types.Scene.main_scene = bpy.props.PointerProperty(type=bpy.types.Scene, name="main scene", description="main_scene_chooser", poll=is_scene_ok)
-    bpy.types.Scene.library_scene = bpy.props.PointerProperty(type=bpy.types.Scene, name="library scene", description="library_scene_picker", poll=is_scene_ok)
-
     # add our addon to the toolbar
     bpy.types.TOPBAR_MT_file_export.append(menu_func_import)
-
-    ## just experiments
-    bpy.types.Scene.main_scenes_list_index = IntProperty(name = "Index for main scenes list", default = 0)
-    bpy.types.Scene.library_scenes_list_index = IntProperty(name = "Index for library scenes list", default = 0)
-
-    #bpy.context.preferences.addons["gltf_auto_export"].preferences.main_scenes_index = 0
-    #bpy.context.preferences.addons["gltf_auto_export"].preferences.library_scenes_index = 0
 
 def unregister():
     for cls in classes:
         bpy.utils.unregister_class(cls)
-
     bpy.types.TOPBAR_MT_file_export.remove(menu_func_import)
-
-    del bpy.types.Scene.main_scene
-    del bpy.types.Scene.library_scene
-
-    del bpy.types.Scene.main_scenes_list_index
-    del bpy.types.Scene.library_scenes_list_index
-
 
 if "gltf_auto_export" == "__main__":
     register()
