+++ tools/gltf_auto_export/auto_export/auto_export.py
@@ -2,19 +2,17 @@ import os
 import bpy
 import traceback
 
-from .helpers_scenes import (get_scenes, )
-from .helpers_collections import (get_collections_in_library, get_exportable_collections, get_collections_per_scene, find_collection_ascendant_target_collection)
-from .helpers_export import (export_main_scene, export_blueprints_from_collections)
-from .helpers import (check_if_blueprints_exist, check_if_blueprint_on_disk)
-from .materials import cleanup_materials, clear_material_info, clear_materials_scene, export_materials, generate_materials_scenes, get_all_materials
-from .scene_components import upsert_scene_components
+from ..helpers_scenes import (get_scenes, )
+from ..helpers_collections import (get_collections_in_library, get_exportable_collections, get_collections_per_scene, find_collection_ascendant_target_collection)
+from ..helpers_export import (export_main_scene, export_blueprints_from_collections)
+from ..helpers import (check_if_blueprints_exist, check_if_blueprint_on_disk)
+from ..materials import cleanup_materials, clear_material_info, clear_materials_scene, export_materials, generate_materials_scenes, get_all_materials
+from ..scene_components import upsert_scene_components
 
-from .config import scene_key
+from ..config import scene_key
 
 """Main function"""
-def auto_export(changes_per_scene, changed_export_parameters):
-    addon_prefs = bpy.context.preferences.addons["gltf_auto_export"].preferences
-
+def auto_export(changes_per_scene, changed_export_parameters, addon_prefs):
     # have the export parameters (not auto export, just gltf export) have changed: if yes (for example switch from glb to gltf, compression or not, animations or not etc), we need to re-export everything
     print ("changed_export_parameters", changed_export_parameters)
     try:
